+++ core/app.ts
@@ -25,9 +25,10 @@ import { enabledModules } from './modules-entry'
 import globalConfig from 'config'
 import { injectReferences } from '@vue-storefront/core/lib/modules'
 import { coreHooksExecutors } from '@vue-storefront/core/hooks'
-import { registerClientModules } from 'src/modules/client';
+import { registerClientModules } from 'src/modules/client'
 import initialStateFactory from '@vue-storefront/core/helpers/initialStateFactory'
-import { createRouter, createRouterProxy } from '@vue-storefront/core/helpers/router';
+import { createRouter, createRouterProxy } from '@vue-storefront/core/helpers/router'
+import areIntlLocalesSupported from 'intl-locales-supported'
 
 const stateFactory = initialStateFactory(store.state)
 
@@ -103,6 +104,15 @@ const createApp = async (ssrContext, config, storeCode = null): Promise<{app: Vu
   registerClientModules()
   registerModules(enabledModules, appContext)
   registerTheme(globalConfig.theme, app, routerProxy, store, globalConfig, ssrContext)
+
+  // Use Intl.js polyfill if needed
+  if (global.Intl) {
+    if (!areIntlLocalesSupported(storeView.i18n.defaultLocale)) {
+      global.Intl = require('intl')
+    }
+  } else {
+    global.Intl = require('intl')
+  }
 
   coreHooksExecutors.afterAppInit()
   // @deprecated from 2.0