+++ core/filters/price.js
@@ -1,9 +1,10 @@
-import { currentStoreView } from '@vue-storefront/core/lib/multistore';
+import { currentStoreView } from '@vue-storefront/core/lib/multistore'
 
 const formatValue = (value, locale) => {
-  const price = Math.abs(parseFloat(value));
-  return price.toLocaleString(locale, { minimumFractionDigits: 2, maximumFractionDigits: 2 });
-};
+  const price = Math.abs(parseFloat(value))
+  const formatter = Intl.NumberFormat(locale, { minimumFractionDigits: 2, maximumFractionDigits: 2 })
+  return formatter.format(price)
+}
 
 const applyCurrencySign = (formattedPrice, { currencySign, priceFormat }) => {
   return priceFormat.replace('{sign}', currencySign).replace('{amount}', formattedPrice)
@@ -15,17 +16,17 @@ const applyCurrencySign = (formattedPrice, { currencySign, priceFormat }) => {
  */
 export function price (value) {
   if (isNaN(value)) {
-    return value;
+    return value
   }
-  const storeView = currentStoreView();
+  const storeView = currentStoreView()
   const { defaultLocale, currencySign, priceFormat } = storeView.i18n
 
-  const formattedValue = formatValue(value, defaultLocale);
+  const formattedValue = formatValue(value, defaultLocale)
   const valueWithSign = applyCurrencySign(formattedValue, { currencySign, priceFormat })
 
   if (value >= 0) {
-    return valueWithSign;
+    return valueWithSign
   } else {
-    return '-' + valueWithSign;
+    return '-' + valueWithSign
   }
 }